Added fuzzy TLS write test
